project/requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.2
gpt4all==1.0.5
greenlet==2.0.2
idna==3.4
pycparser==2.21
pycryptodome==3.18.0
PyMySQL==1.1.0
requests==2.31.0
SQLAlchemy==2.0.19
SQLAlchemy-Utils==0.41.1
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.3
